+++ CitizenImporter/CitizenImporter.cs
@@ -7,6 +7,11 @@ using HarmonyLib;
 using Cpp2IL.Core.Extensions;
 
 using static Descriptors;
+using System.Linq;
+using UniverseLib;
+
+
+
 
 #if MONO
 using BepInEx.Unity.Mono;
@@ -32,6 +37,11 @@ namespace CitizenImporter
         // People from file loading
         class HumanOverride
         {
+            public int id;
+            public static int idCounter;
+            public static Dictionary<int, HumanOverride> createdHumans;
+            public int gameId = -1;
+
             public string firstName;
             public string lastName;
             public string casualName;
@@ -42,24 +52,44 @@ namespace CitizenImporter
 
             public string gender;
             public string attractedTo;
-            
+
             public string ethnicity1;
             public string ethnicity2;
 
             public string height;
             public string buildType;
             public string hairColour;
-            public string hairColourExact;
+            public string hairColourExact = "";
             public string hairType;
             public string eyeColour;
 
             public string glasses;
             public string beard;
             public string moustache;
+
+            public string partner;
+
+            public HumanOverride()
+            {
+                id = ++idCounter;
+                createdHumans[id] = this;
+            }
+
+            // Human line numbers, excluding the heading row
+            public static void ResetIdCounter()
+            {
+                idCounter = 1;
+                createdHumans = new Dictionary<int, HumanOverride>();
+            }
         }
 
-        static System.Collections.Generic.List<HumanOverride> loadedHumanOverrides = new System.Collections.Generic.List<HumanOverride>();
-        static System.Collections.Generic.Dictionary<int, (Human, HumanOverride)> overridenHumans = new System.Collections.Generic.Dictionary<int, (Human, HumanOverride)>();
+        static Dictionary<Human.Gender, List<HumanOverride>> loadedHumanOverrides = new Dictionary<Human.Gender, List<HumanOverride>>()
+        {
+            { Human.Gender.male, new List<HumanOverride>() },
+            { Human.Gender.female, new List<HumanOverride>() },
+            { Human.Gender.nonBinary, new List<HumanOverride>() }
+        };
+        static Dictionary<int, (Human, HumanOverride)> overridenHumans = new Dictionary<int, (Human, HumanOverride)>();
 
 #if MONO
         private void Awake()
@@ -79,72 +109,226 @@ namespace CitizenImporter
             PluginLogger.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} is patched!");
         }
 
-        static void ReloadCitizensFromFile()
+        static bool ReloadCitizensFromFile()
         {
             var path = Path.Combine(Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location), "citizens.csv");
             if (!File.Exists(path))
             {
                 PluginLogger.LogError("Citizen file not found, no citizen overrides will be loaded. See https://docs.google.com/spreadsheets/d/11OvB572ezm2j-iO2rd89XHInrTNXey83eLdw2VJcujE/edit#gid=1379962932");
-                return;
+                return false;
             }
 
             string[] citizensCSV = File.ReadAllLines(path);
 
-            loadedHumanOverrides.Clear();
+            loadedHumanOverrides[Human.Gender.male].Clear();
+            loadedHumanOverrides[Human.Gender.female].Clear();
+            loadedHumanOverrides[Human.Gender.nonBinary].Clear();
+
+            HumanOverride.ResetIdCounter();
 
             // Header
             for (int i = 1; i < citizensCSV.Length; i++)
             {
                 string[] vals = citizensCSV[i].Split(",");
-                loadedHumanOverrides.Add(new HumanOverride()
+                var newHumanOverride = new HumanOverride()
                 {
-                    firstName = vals[0],
-                    lastName = vals[1],
-                    casualName = vals[2],
-                    dayOfBirth = vals[3],
-                    monthOfBirth = vals[4],
-                    age = vals[5],
-                    gender = vals[6],
-                    attractedTo = vals[7],
-                    ethnicity1 = vals[8],
-                    ethnicity2 = vals[9],
-                    height = vals[10],
-                    buildType = vals[11],
-                    hairColour = vals[12],
-                    hairType = vals[13],
-                    eyeColour = vals[14],
-                    glasses = vals[15],
-                    beard = vals[16],
-                    moustache = vals[17]
-                });
+                    firstName = vals.ElementAtOrDefault(0),
+                    lastName = vals.ElementAtOrDefault(1),
+                    casualName = vals.ElementAtOrDefault(2),
+                    dayOfBirth = vals.ElementAtOrDefault(3),
+                    monthOfBirth = vals.ElementAtOrDefault(4),
+                    age = vals.ElementAtOrDefault(5),
+                    gender = vals.ElementAtOrDefault(6),
+                    attractedTo = vals.ElementAtOrDefault(7),
+                    ethnicity1 = vals.ElementAtOrDefault(8),
+                    ethnicity2 = vals.ElementAtOrDefault(9),
+                    height = vals.ElementAtOrDefault(10),
+                    buildType = vals.ElementAtOrDefault(11),
+                    hairColour = vals.ElementAtOrDefault(12),
+                    hairType = vals.ElementAtOrDefault(13),
+                    eyeColour = vals.ElementAtOrDefault(14),
+                    glasses = vals.ElementAtOrDefault(15),
+                    beard = vals.ElementAtOrDefault(16),
+                    moustache = vals.ElementAtOrDefault(17),
+                    partner = vals.ElementAtOrDefault(18)
+                };
+
+                // Don't use non-binary here, because it's just too rare - citizens likely won't generate 
+                Human.Gender genderToAssign = (Human.Gender)Random.Range(0, 2);
+                if (newHumanOverride.gender.Length != 0)
+                {
+                    if (!System.Enum.TryParse<Human.Gender>(newHumanOverride.gender, true, out genderToAssign))
+                    {
+                        LogParseFailure(null, newHumanOverride, "gender");
+                    }
+                }
+
+                loadedHumanOverrides[genderToAssign].Add(newHumanOverride);
             }
 
-            if (loadedHumanOverrides.Count == 0)
+            // Probably don't need to shuffle the order, the game seems to create citizens in a fairly random order?
+
+            if (loadedHumanOverrides.Aggregate(0, (total, tuple) => total + tuple.Value.Count) == 0)
             {
                 PluginLogger.LogError("Citizen file found, but no citizen overrides loaded. No overrides will be applied.");
-                return;
+                return false;
+            }
+            return true;
+        }
+
+        [HarmonyPatch(typeof(CitizenCreator), "Populate")]
+        public class CitizenCreator_Populate
+        {
+            public static void Prefix()
+            {
+                PluginLogger.LogInfo("Reloading citizen overrides");
+                Creator_SetComplete.citizensLoadedFromFile = ReloadCitizensFromFile();
+                Creator_SetComplete.unusedPartnersCount = 0;
             }
         }
 
-        static void LogParseFailure(Human human, HumanOverride humanOverride, string field)
+        [HarmonyPatch(typeof(Creator), "SetComplete")]
+        public class Creator_SetComplete
         {
-            PluginLogger.LogWarning($"Human {human.humanID} (Override name: {humanOverride.firstName} {humanOverride.lastName}) failed to parse {field}");
+            public static bool citizensLoadedFromFile = false;
+            public static int unusedPartnersCount = 0;
+            public static void Postfix(Creator __instance)
+            {
+                if (citizensLoadedFromFile && __instance.GetActualType() == typeof(CitizenCreator))
+                {
+                    PluginLogger.LogWarning($"Citizen overrides applied. {loadedHumanOverrides.Aggregate(0, (total, tuple) => total + tuple.Value.Count) + unusedPartnersCount} overrides were not used ({Creator_SetComplete.unusedPartnersCount} were unused partners)");
+                    /*
+                    foreach (var keyValue in overridenHumans)
+                    {
+                        var h = keyValue.Value.Item1;
+                        PluginLogger.LogInfo($"Citizen override: {h.citizenName} ({h.gender}, attracted to {h.attractedTo.ToArray().Join()}). {(h.partner != null ? ("Partner: " + h.partner.citizenName) : "")}");
+                    }
+
+                    PluginLogger.LogWarning($"Citizens not attracted to partners:");
+
+                    foreach(var cit in CityData.Instance.citizenDirectory)
+                    {
+                        if (cit.partner != null)
+                        {
+                            if (!cit.attractedTo.Contains(cit.partner.gender))
+                            {
+                                PluginLogger.LogInfo($"{cit.citizenName} ({cit.gender}, attracted to {cit.attractedTo.ToArray().Join()}) is not attracted to partner {cit.partner.citizenName} ({cit.partner.gender})");
+                            }
+                        }
+                    }
+                    */
+                }
+            }
         }
 
-        // Partner is set after this, random dice roll.. might have to override it to force, no neat hook location
+        static void LogParseFailure(Human? human, HumanOverride humanOverride, string field)
+        {
+            PluginLogger.LogWarning($"Human {human?.humanID} (Override name: {humanOverride.firstName} {humanOverride.lastName}) failed to parse {field}");
+        }
+
+        // Partner is set after this, random dice roll.. How can we force them to spawn if requested?
         [HarmonyPatch(typeof(Human), "SetSexualityAndGender")]
         public class Human_SetSexualityAndGender
         {
-            public static void Postfix(Human __instance)
+            public static void Postfix(ref Human __instance)
             {
                 if (SessionData.Instance.isFloorEdit || CityConstructor.Instance.generateNew)
                 {
                     if (loadedHumanOverrides.Count > 0)
                     {
-                        if (__instance.gender == Human.Gender.male && __instance.attractedTo.Contains(Human.Gender.female) && __instance.attractedTo.Count == 1)
+                        var nextSuitableHuman = 0;
+                        for (; nextSuitableHuman < loadedHumanOverrides[__instance.gender].Count; nextSuitableHuman++)
                         {
-                            overridenHumans[__instance.humanID] = (__instance, loadedHumanOverrides.RemoveAndReturn(0));
-                            PluginLogger.LogInfo($"Setting override: {overridenHumans[__instance.humanID].Item2.firstName} {overridenHumans[__instance.humanID].Item2.lastName} will replace {__instance.humanID}");
+                            var testingHuman = loadedHumanOverrides[__instance.gender][nextSuitableHuman];
+
+                            // Make sure that 'partner' relationships are one-way, to prevent loops
+                            if (int.TryParse(testingHuman.partner, out int partnerId))
+                            {
+                                if (HumanOverride.createdHumans[partnerId].gameId != -1)
+                                {
+                                    // If we already created this person's partner, they can't spawn
+                                    loadedHumanOverrides[__instance.gender].RemoveAt(nextSuitableHuman);
+                                    Creator_SetComplete.unusedPartnersCount++;
+                                    continue;
+                                }
+                            }
+
+                            if(System.Enum.TryParse(HumanOverride.createdHumans[partnerId].gender, true, out Human.Gender partnerGender))
+                            {
+                                // Our partner has a fixed gender
+                                if (__instance.attractedTo.Contains(partnerGender))
+                                {
+                                    __instance.attractedTo.Clear();
+                                    __instance.attractedTo.Add(partnerGender);
+                                    break;
+                                }
+                            }
+                            else if (System.Enum.TryParse(testingHuman.attractedTo, true, out Human.Gender attractedTo))
+                            {
+                                // We have an assigned attracted to, make sure it matches
+                                if (__instance.attractedTo.Contains(attractedTo))
+                                {
+                                    __instance.attractedTo.Clear();
+                                    __instance.attractedTo.Add(attractedTo);
+                                    break;
+                                }
+                            }
+                            else
+                            {
+                                // If we don't have an assigned attracted to value, we can use this person regardless
+                                break;
+                            }
+                        }
+
+                        if (nextSuitableHuman < loadedHumanOverrides[__instance.gender].Count)
+                        {
+                            overridenHumans[__instance.humanID] = (__instance, loadedHumanOverrides[__instance.gender].RemoveAndReturn(nextSuitableHuman));
+                            overridenHumans[__instance.humanID].Item2.gameId = __instance.humanID;
+                            // Nonbinary, non specified, will cause us issues later, so remove it
+                            if (__instance.attractedTo.Count > 1)
+                                __instance.attractedTo.Remove(Human.Gender.nonBinary);
+                            PluginLogger.LogInfo($"Setting override: {overridenHumans[__instance.humanID].Item2.firstName} {overridenHumans[__instance.humanID].Item2.lastName} ({overridenHumans[__instance.humanID].Item2.id}) will replace {__instance.humanID}");
+                        }
+                    }
+                }
+            }
+        }
+
+        [HarmonyPatch(typeof(Human), "GenerateSuitableGenderAndSexualityForParnter")]
+        public class Human_GenerateSuitableGenderAndSexualityForParnter
+        {
+            public static void Postfix(Human __instance, Citizen newPartner)
+            {
+                if (SessionData.Instance.isFloorEdit || CityConstructor.Instance.generateNew)
+                {
+                    if (overridenHumans.ContainsKey(newPartner.humanID) && int.TryParse(overridenHumans[newPartner.humanID].Item2.partner, out var partnerHumanId))
+                    {
+                        int partnerId = -1;
+                        Human.Gender foundAttractedTo = __instance.gender;
+                        foreach (var gender in new Human.Gender[] { Human.Gender.male, Human.Gender.female, Human.Gender.nonBinary })
+                        {
+                            partnerId = loadedHumanOverrides[gender].FindIndex(ho => ho.id == partnerHumanId);
+                            foundAttractedTo = gender;
+                            if (partnerId > -1) break;
+                        }
+
+                        if (partnerId >= 0)
+                        {
+                            var partnerOverride = loadedHumanOverrides[foundAttractedTo].RemoveAndReturn(partnerId);
+                            if (partnerOverride.gender == "" || (System.Enum.TryParse(partnerOverride.gender, true, out Human.Gender partnerGender) && newPartner.attractedTo.Contains(partnerGender)))
+                            {
+                                overridenHumans[__instance.humanID] = (__instance, partnerOverride);
+                                overridenHumans[__instance.humanID].Item2.gameId = __instance.humanID;
+                                PluginLogger.LogInfo($"Setting override: {overridenHumans[__instance.humanID].Item2.firstName} {overridenHumans[__instance.humanID].Item2.lastName} will replace {__instance.humanID} as a partner of {newPartner.humanID}");
+                            }
+                            else
+                            {
+                                PluginLogger.LogInfo($"Invalid partner for {partnerHumanId} (Gender)");// , needed {} found {})");
+                            }
+                        }
+                        else
+                        {
+                            PluginLogger.LogInfo($"Invalid partner for {partnerHumanId}");
                         }
                     }
                 }
@@ -159,8 +343,8 @@ namespace CitizenImporter
                 if (overridenHumans.ContainsKey(__instance.humanID))
                 {
                     var overrideHuman = overridenHumans[__instance.humanID].Item2;
-                    if(
-                        int.TryParse(overrideHuman.dayOfBirth, out int dayOfBirth) &&  dayOfBirth > 0 && dayOfBirth <= 31 && 
+                    if (
+                        int.TryParse(overrideHuman.dayOfBirth, out int dayOfBirth) && dayOfBirth > 0 && dayOfBirth <= 31 &&
                         int.TryParse(overrideHuman.monthOfBirth, out int monthOfBirth) && monthOfBirth > 0 && monthOfBirth <= 12 &&
                         int.TryParse(overrideHuman.age, out int age)
                     )
@@ -223,12 +407,12 @@ namespace CitizenImporter
                     {
                         if (overrideHuman.ethnicity2 == null || overrideHuman.ethnicity2 == "") overrideHuman.ethnicity2 = overrideHuman.ethnicity1;
 
-                        if (!System.Enum.TryParse<EthnicGroup>(overrideHuman.ethnicity1, out var eth1))
+                        if (!System.Enum.TryParse<EthnicGroup>(overrideHuman.ethnicity1, true, out var eth1))
                         {
                             LogParseFailure(__instance.citizen, overrideHuman, "ethnicity1");
                             return;
                         }
-                        if (!System.Enum.TryParse<EthnicGroup>(overrideHuman.ethnicity2, out var eth2))
+                        if (!System.Enum.TryParse<EthnicGroup>(overrideHuman.ethnicity2, true, out var eth2))
                         {
                             LogParseFailure(__instance.citizen, overrideHuman, "ethnicity2");
                             return;
@@ -262,7 +446,7 @@ namespace CitizenImporter
                     var overrideHuman = overridenHumans[__instance.citizen.humanID].Item2;
                     if (overrideHuman.eyeColour != null && overrideHuman.eyeColour != "")
                     {
-                        if (System.Enum.TryParse<EyeColour>(overrideHuman.eyeColour, out var setEyeColour))
+                        if (System.Enum.TryParse<EyeColour>(overrideHuman.eyeColour, true, out var setEyeColour))
                         {
                             __instance.eyeColour = setEyeColour;
                         }
@@ -283,9 +467,9 @@ namespace CitizenImporter
                 if (overridenHumans.ContainsKey(__instance.citizen.humanID))
                 {
                     var overrideHuman = overridenHumans[__instance.citizen.humanID].Item2;
-                    __instance.citizen.firstName = overrideHuman.firstName;
-                    __instance.citizen.surName = overrideHuman.lastName;
-                    __instance.citizen.casualName = overrideHuman.casualName ?? __instance.citizen.firstName;
+                    __instance.citizen.firstName = overrideHuman.firstName.Length > 0 ? overrideHuman.firstName : __instance.citizen.firstName;
+                    __instance.citizen.surName = overrideHuman.lastName.Length > 0 ? overrideHuman.lastName : __instance.citizen.surName;
+                    __instance.citizen.casualName = overrideHuman.casualName.Length > 0 ? overrideHuman.casualName : __instance.citizen.firstName;
                     __instance.citizen.citizenName = __instance.citizen.firstName + " " + __instance.citizen.surName;
                     __instance.citizen.gameObject.name = __instance.citizen.citizenName;
                     __instance.citizen.name = __instance.citizen.citizenName;
@@ -301,13 +485,13 @@ namespace CitizenImporter
                 if (overridenHumans.ContainsKey(__instance.citizen.humanID))
                 {
                     var overrideHuman = overridenHumans[__instance.citizen.humanID].Item2;
-                    if (overrideHuman.hairColour != null && overrideHuman.hairColour != "")
+                    if (overrideHuman.hairColour != "")
                     {
-                        if (System.Enum.TryParse<HairColour>(overrideHuman.hairColour, out var setHairColour))
+                        if (System.Enum.TryParse<HairColour>(overrideHuman.hairColour, true, out var setHairColour))
                         {
                             __instance.hairColourCategory = setHairColour;
 
-                            if (overrideHuman.hairColourExact != null && overrideHuman.hairColourExact != "")
+                            if (overrideHuman.hairColourExact != "")
                             {
                                 if (ColorUtility.TryParseHtmlString(overrideHuman.hairColourExact, out var setExactHairColour))
                                 {
@@ -336,9 +520,9 @@ namespace CitizenImporter
                         }
                     }
 
-                    if (overrideHuman.hairType != null)
+                    if (overrideHuman.hairType != "")
                     {
-                        if (System.Enum.TryParse<HairStyle>(overrideHuman.hairType, out var setHairStyle))
+                        if (System.Enum.TryParse<HairStyle>(overrideHuman.hairType, true, out var setHairStyle))
                         {
                             __instance.hairType = setHairStyle;
                         }
@@ -359,9 +543,9 @@ namespace CitizenImporter
                 if (overridenHumans.ContainsKey(__instance.citizen.humanID))
                 {
                     var overrideHuman = overridenHumans[__instance.citizen.humanID].Item2;
-                    if (overrideHuman.buildType != null && overrideHuman.buildType != "")
+                    if (overrideHuman.buildType != "")
                     {
-                        if (System.Enum.TryParse<BuildType>(overrideHuman.buildType, out var setValue))
+                        if (System.Enum.TryParse<BuildType>(overrideHuman.buildType, true, out var setValue))
                         {
                             __instance.build = setValue;
                         }
@@ -372,7 +556,7 @@ namespace CitizenImporter
                     }
 
                     // Setup the height override if set
-                    if (overrideHuman.height != null && overrideHuman.height != "" && float.TryParse(overrideHuman.height, out var height))
+                    if (overrideHuman.height != "" && float.TryParse(overrideHuman.height, out var height))
                     {
                         __instance.citizen.seed = $"piecitizenoverride_{height}_{__instance.citizen.seed}";
                     }
@@ -405,7 +589,7 @@ namespace CitizenImporter
 
                     if (overrideHuman.glasses != null && overrideHuman.glasses != "")
                     {
-                        if(overrideHuman.glasses == "true") glasses = OverrideFeature.On;
+                        if (overrideHuman.glasses == "true") glasses = OverrideFeature.On;
                         else if (overrideHuman.glasses == "false") glasses = OverrideFeature.Off;
                     }
 
@@ -413,8 +597,8 @@ namespace CitizenImporter
                     {
                         if (overrideHuman.beard == "true") beard = OverrideFeature.On;
                         else if (overrideHuman.beard == "false") beard = OverrideFeature.Off;
-                    }            
-                    
+                    }
+
                     if (overrideHuman.moustache != null && overrideHuman.moustache != "")
                     {
                         if (overrideHuman.moustache == "true") moustache = OverrideFeature.On;