+++ CitizenImporter/ModFolderContent/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.2
+
+* Any names (first, last, both) that aren't specified will now use the game's randomisation
+* Partners can now be specified based on row number in the CSV (Make sure the AttractedTo value matches the partners Gender value)
+
+# 0.1.1
+
+* Fixed a bug with assigning overrides to other genders and orientations
+
 # 0.1.0
  
 * Initial release